Split GPS tests
